+++ src/components/Search.tsx
@@ -1,0 +1,4 @@
+export const Search = ()  => {
+
+    return<>Search</>
+}
