+++ src/pages/Home.tsx
@@ -1,3 +1,10 @@
+import { JobList } from "../components/JobList"
+import { Search } from "../components/Search"
+
 export const Home = () => {
-    return<>Home</>
+    return<>
+    Home
+        <Search/>
+        <JobList/>
+    </>
 }
