--- src/pages/Job.tsx
@@ -1,3 +1,0 @@
-export const Job = () => {
-    return<>Job</>
-}
